+++ R/cpi_u.R
@@ -1,6 +1,6 @@
 #' Consumer Price Index for Urban Consumers (CPI-U) data
 #'
-#' Monthly or annual CPI-U price indexes
+#' Monthly, quarterly, or annual CPI-U price indexes
 #'
 #' @docType data
 #' @name cpi_u
@@ -15,6 +15,15 @@
 #'   \item{cpi_u}{value of the CPI-U price index}
 #' }
 #'
+#' #' @format The data frames `cpi_u_quarterly_sa` and `cpi_u_quarterly_nsa`
+#' contain, respectively, seasonally adjusted or not seasonally adjusted quarterly
+#' price index levels of the CPI-U. They have three columns:
+#' \describe{
+#'   \item{year}{numeric year}
+#'   \item{quarter}{numeric calendar quarter (1-4)}
+#'   \item{cpi_u}{value of the CPI-U price index}
+#' }
+#'
 #' The data frame `cpi_u_annual` contains the annual
 #' price index level of the CPI-U. It has two columns:
 #' \describe{
@@ -26,7 +35,10 @@
 #' @seealso [`cpi_u_rs_monthly_nsa`][cpi_u_rs] and [`cpi_u_rs_annual`][cpi_u_rs]
 #' for the CPI-U-RS, an index that is more consistent over time than the CPI-U
 #' and incorporates more recent improvements to the CPI-U into the
-#' entire time series.
+#' entire time series. The quarterly `cpi_u_quarterly_nsa` and
+#' `cpi_u_quarterly_sa` series are the quarterly means of their
+#' respective monthly series.
+#'
 #' @examples
 #' cpi_u_monthly_nsa
 
@@ -37,6 +49,14 @@
 #' @rdname cpi_u
 #' @format NULL
 "cpi_u_monthly_sa"
+
+#' @rdname cpi_u
+#' @format NULL
+"cpi_u_quarterly_nsa"
+
+#' @rdname cpi_u
+#' @format NULL
+"cpi_u_quarterly_sa"
 
 #' @rdname cpi_u
 #' @format NULL