+++ R/cpi_u_x1.R
@@ -23,10 +23,8 @@
 #' }
 #'
 #' @source Emailed spreadsheet from the US Bureau of Labor Statistics
-#' @seealso [`cpi_u_rs_extended_monthly_nsa`][cpi_u_rs_extended] and [`cpi_u_extended_rs_annual`][cpi_u_rs_extended]
-#' for an extended version of the CPI-U-RS, using the [`cpi_u`] and [`cpi_u_x1`]
-#' which is more consistent over time than the CPI-U and incorporates more recent
-#' improvements to the CPI-U into the entire time series.
+#' @seealso [`c_cpi_u_extended_annual`][`c_cpi_u_extended`] extends the
+#' CPI-U-X1 from 1937 to the present using the [`c_cpi_u`] and other series.
 #' @examples
 #' cpi_u_x1_annual