+++ data-raw/_targets.R
@@ -2,7 +2,7 @@
 source("data-raw/packages.R")
 
 ## Globals
-download_date <- ymd("2024-04-11")
+download_date <- ymd("2024-05-29")
 
 ## Functions
 lapply(list.files("./data-raw/R", full.names = TRUE), source)
@@ -16,6 +16,7 @@ tar_plan(
   tar_file(cpi_u_nsa_raw, fetch_cpi_u_nsa(download_date)),
   tar_file(c_cpi_u_raw, fetch_c_cpi_u(download_date)),
   tar_file(pce_monthly_raw, fetch_pce_monthly_sa(download_date)),
+  tar_file(pce_quarterly_raw, fetch_pce_quarterly_sa(download_date)),
   tar_file(pce_annual_raw, fetch_pce_annual(download_date)),
 
   # final data: csv and package data
@@ -23,13 +24,10 @@ tar_plan(
   tar_file(cpi_u_rs_data, create_cpi_u_rs(cpi_u_rs_raw)),
   tar_file(c_cpi_u_data, create_c_cpi_u(c_cpi_u_raw)),
   tar_file(cpi_u_data, create_cpi_u(cpi_u_nsa_raw, cpi_u_sa_raw)),
-  tar_file(cpi_u_rs_extended_data, create_cpi_u_rs_extended(
-    cpi_u_data, cpi_u_x1_data, cpi_u_rs_data
-  )),
   tar_file(c_cpi_u_extended_data, create_c_cpi_u_extended(
     cpi_u_data, cpi_u_x1_data, cpi_u_rs_data, c_cpi_u_data
   )),
-  tar_file(pce_data, create_pce(pce_monthly_raw, pce_annual_raw)),
+  tar_file(pce_data, create_pce(pce_monthly_raw, pce_quarterly_raw, pce_annual_raw)),
   tar_target(fed_mw_monthly_data, create_fed_mw_monthly(us_minimum_wage_raw)),
   tar_file(fed_mw_annual_data, create_fed_mw_annual(fed_mw_monthly_data)),
 
@@ -40,7 +38,6 @@ tar_plan(
     c_cpi_u_extended_data,
     cpi_u_x1_data,
     cpi_u_rs_data,
-    cpi_u_rs_extended_data,
     pce_data
   ))
 )