+++ NEWS.md
@@ -1,3 +1,7 @@
+## Version 0.11.0
+* added quarterly versions of CPI-U, C-CPI-U, extended C-CPI-U, and PCE 
+* updated series through April 2024 when available
+
 ## Version 0.10.0
 * updated series through March 2024 when available